# deep_research/api.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportListItem": {
        "description": "Compact report card shown in the list view.",
        "properties": {
          "citation_count": {
            "default": 0,
            "title": "Citation Count",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "has_pdf": {
            "default": false,
            "title": "Has Pdf",
            "type": "boolean"
          },
          "iterations": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iterations"
          },
          "markdown_length": {
            "default": 0,
            "title": "Markdown Length",
            "type": "integer"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "snippet": {
            "default": "",
            "title": "Snippet",
            "type": "string"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "started_at",
          "query",
          "path"
        ],
        "title": "ReportListItem",
        "type": "object"
      },
...
      "ResearchJobStatus": {
        "properties": {
          "archived": {
            "default": false,
            "title": "Archived",
            "type": "boolean"
          },
          "attach_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attach To"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "paused_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paused At"
          },
          "phase": {
            "default": "",
            "title": "Phase",
            "type": "string"
          },
          "query": {
            "default": "",
            "title": "Query",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "started_at": {
            "default": 0.0,
            "title": "Started At",
            "type": "number"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "step": {
            "default": "",
            "title": "Step",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "ResearchJobStatus",
        "type": "object"
      },
      "ResearchStartRequest": {
        "properties": {
          "attach_to_run_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attach To Run Id"
          },
          "path_override": {
            "anyOf": [
              {
                "enum": [
                  "quick",
                  "deep",
                  "academic",
                  "url_source"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Override"
          },
          "query": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "webhook_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          }
        },
        "required": [
          "query"
        ],
        "title": "ResearchStartRequest",
        "type": "object"
      },
      "ResearchStartResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "ResearchStartResponse",
        "type": "object"
      },
      "SearchHitItem": {
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "authors": {
            "default": "",
            "title": "Authors",
            "type": "string"
          },
          "score": {
            "default": 0.0,
            "title": "Score",
            "type": "number"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "artifact_id"
        ],
        "title": "SearchHitItem",
        "type": "object"
      },
      "SearchResponse": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/SearchHitItem"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "q": {
            "title": "Q",
            "type": "string"
          },
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportListItem"
            },
            "title": "Reports",
            "type": "array"
          }
        },
        "required": [
          "q"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Deep Research Library Web UI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/research": {
      "post": {
        "operationId": "start_research_api_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Research",
        "tags": [
          "research"
        ]
      }
    },
...
    "/api/research/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_status_api_research_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchJobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Status",
        "tags": [
          "research"
        ]
      }
    },
    "/api/research/jobs/{job_id}/abandon": {
      "post": {
        "operationId": "abandon_job_api_research_jobs__job_id__abandon_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchJobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Abandon Job",
        "tags": [
          "research"
        ]
      }
    },
    "/api/research/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_research_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchJobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "research"
        ]
      }
    },
    "/api/research/jobs/{job_id}/pause": {
      "post": {
        "operationId": "pause_job_api_research_jobs__job_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchJobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Job",
        "tags": [
          "research"
        ]
      }
    },
    "/api/research/jobs/{job_id}/resume": {
      "post": {
        "operationId": "resume_job_api_research_jobs__job_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchJobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Job",
        "tags": [
          "research"
        ]
      }
    },
    "/api/research/jobs/{job_id}/stream": {
      "get": {
        "operationId": "stream_job_api_research_jobs__job_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Job",
        "tags": [
          "research"
        ]
      }
    },
    "/api/search": {
      "get": {
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "library"
        ]
      }
    },
...
  }
}